+++ electron/tsconfig.json
@@ -7,7 +7,6 @@
     "outDir": "../build",
     "rootDir": "../",
     "baseUrl": "../",
-    "jsx": "react-jsx",
     "noEmitOnError": true,
     "resolveJsonModule": true
   },
